+++ tests/PolynomialTest.java
@@ -1,10 +1,23 @@
+import show.ezkz.casprzak.core.config.Settings;
 import show.ezkz.casprzak.core.functions.GeneralFunction;
 import org.junit.jupiter.api.Test;
+import show.ezkz.casprzak.core.functions.binary.Pow;
+import show.ezkz.casprzak.core.functions.commutative.Product;
+import show.ezkz.casprzak.core.functions.commutative.Sum;
+import show.ezkz.casprzak.core.functions.endpoint.Constant;
+import show.ezkz.casprzak.core.functions.endpoint.Variable;
+import show.ezkz.casprzak.core.functions.unitary.specialcases.Exp;
+import show.ezkz.casprzak.core.tools.defaults.DefaultFunctions;
+import show.ezkz.casprzak.core.tools.functiongenerators.HermitePolynomial;
+import show.ezkz.casprzak.core.tools.functiongenerators.LaguerrePolynomial;
+import show.ezkz.casprzak.core.tools.functiongenerators.LegrendePolynomial;
+import show.ezkz.casprzak.core.tools.functiongenerators.TaylorSeries;
 import show.ezkz.casprzak.parsing.FunctionParser;
 import show.ezkz.casprzak.core.tools.PolynomialTools;
 
-import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static show.ezkz.casprzak.core.tools.defaults.DefaultFunctions.*;
 
 public class PolynomialTest {
     @Test
@@ -54,5 +67,56 @@ public class PolynomialTest {
         GeneralFunction test = FunctionParser.parseSimplified("5x^3+x^2+5x+3269485");
         assertTrue(PolynomialTools.isPolynomial(test));
     }
+
+    @Test
+	void legendrePolynomial() {
+    	Variable test1 = new Variable(Settings.singleVariableDefault);
+    	assertEquals(test1, LegrendePolynomial.legrendePolynomial(1));
+		test1 = new Variable("j");
+		assertEquals(test1, LegrendePolynomial.legrendePolynomial(1, "j"));
+		assertEquals(new Constant(1/Math.sqrt(2)), LegrendePolynomial.normalLegrendePolynomial(0));
+		assertEquals(new Product(new Constant(Math.sqrt(3/2.)), new Variable("t")), LegrendePolynomial.normalLegrendePolynomial(1, "t"));
+		assertEquals(new Constant(Math.sqrt(2/5.)), LegrendePolynomial.normalizingConstant(2));
+
+	}
+
+	@Test
+	void hermitePolynomial() {
+		Variable test1 = new Variable(Settings.singleVariableDefault);
+		assertEquals(new Product(TWO, test1), HermitePolynomial.hermitePolynomial(1));
+		test1 = new Variable("j");
+		assertEquals(new Product(TWO, test1), HermitePolynomial.hermitePolynomial(1, "j"));
+		assertEquals(new Pow(new Constant(-.25), PI), HermitePolynomial.normalHermitePolynomial(0));
+		assertEquals(new Product(new Constant(Math.sqrt(2)), new Pow(new Constant(-.25), PI), new Variable("t")), HermitePolynomial.normalHermitePolynomial(1, "t"));
+		assertEquals(new Pow(new Constant(.25), PI), HermitePolynomial.normalizingConstant(0));
+
+	}
+
+	@Test
+	void laguerrePolynomial() {
+		Variable test1 = new Variable(Settings.singleVariableDefault);
+		assertEquals(new Sum(ONE, negative(test1)), LaguerrePolynomial.laguerrePolynomial(1));
+		test1 = new Variable("j");
+		assertEquals(new Sum(ONE, negative(test1)), LaguerrePolynomial.laguerrePolynomial(1, "j"));
+		assertEquals(ONE, LaguerrePolynomial.laguerrePolynomial(0));
+		assertEquals(FunctionParser.parseSimplified("(1/2 x^2-2x+1)"), LaguerrePolynomial.laguerrePolynomial(2));
+		assertEquals(FunctionParser.parseSimplified("(-1/6 x^3+ 3/2 x^2-3x+1)"), LaguerrePolynomial.laguerrePolynomial(3));
+
+	}
+
+	@Test
+	void taylorSeriesVariable1() {
+		Variable test1 = new Variable(Settings.singleVariableDefault);
+		assertEquals(new Sum(ONE, test1), TaylorSeries.makeTaylorSeries(new Exp(test1), 1));
+		test1 = new Variable("j");
+		assertEquals(new Sum(ONE, test1), TaylorSeries.makeTaylorSeries(new Exp(test1), 1));
+	}
+
+	@Test
+	void taylorSeriesVariable2() {
+		Variable x = new Variable(Settings.singleVariableDefault);
+		Variable y = new Variable("y");
+		assertThrows(UnsupportedOperationException.class, () -> TaylorSeries.makeTaylorSeries(new Exp(new Sum(x, y)), 1));
+	}
 
 }